refactor(content): extract EmptyMessage into its own file
