+++ apps/mesh/src/web/components/sandbox/content/content-browser.tsx
@@ -153,6 +153,7 @@ import {
 import { PageJsonDialog } from "@/web/components/sections-editor/page-json-dialog";
 import { RunnableBlocksBrowser } from "./runnable-blocks-browser";
 import { countAvailableRunnables } from "./runnable-catalog";
+import { EmptyMessage } from "./empty-message";
 import { useBlocksPreviewWorkspace } from "@/web/components/sandbox/blocks/blocks-preview-workspace-context";
 import type { BlocksTarget } from "@/web/components/sandbox/blocks/blocks-preview-workspace-state";
 
@@ -3032,26 +3033,4 @@ function SandboxStateRenderer({
         />
       );
   }
-}
-
-export function EmptyMessage({
-  icon: Icon,
-  title,
-  description,
-}: {
-  icon?: React.ComponentType<{ size?: number; className?: string }>;
-  title: string;
-  description?: string;
-}) {
-  return (
-    <div className="flex h-full w-full flex-col items-center justify-center gap-2 p-6 text-center text-sm text-muted-foreground">
-      {Icon && <Icon size={24} className="text-muted-foreground/60" />}
-      <div>{title}</div>
-      {description && (
-        <div className="text-xs text-muted-foreground/80 max-w-sm">
-          {description}
-        </div>
-      )}
-    </div>
-  );
 }
